+++ src/hooks/Input/useExpiredDateInput.js
@@ -4,8 +4,8 @@ import { numberRegex } from 'constant/regularExpression';
 import { CardContext } from 'context/CardContext';
 import { INPUT_ACTION } from 'Reducer/InputtedInfoReducer';
 import validator from 'validation';
-import useFocus from 'hooks/useFocus';
-import useSomeInput from 'hooks/useSomeInput';
+import useFocus from 'hooks/Input/useFocus';
+import useSomeInput from 'hooks/Input/useSomeInput';
 
 function useExpiredDateInput(inputNames) {
   const { inputtedInfoDispatch } = useContext(CardContext);
